third_party/WebKit/LayoutTests/mojo/connection.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script>
'use strict';

setup({ explicit_done: true });

define([
    'mojo/public/js/bindings',
    'mojo/public/js/core',
    'mojo/public/interfaces/bindings/tests/sample_interfaces.mojom',
    'mojo/public/interfaces/bindings/tests/sample_service.mojom',
], function(bindings, core, sample_interfaces, sample_service) {

  promise_test(async () => {
    function ServiceImpl() {
    }

    ServiceImpl.prototype.frobinate = function(foo, baz, port) {
      assert_equals(foo.name, "Example name");
      assert_equals(baz, sample_service.Service.BazOptions.REGULAR);
      assert_true(port.ptr.isBound());
      port.ptr.reset();

      return Promise.resolve({result: 42});
    };

    var service = new sample_service.ServicePtr();
    var serviceBinding = new bindings.Binding(sample_service.Service,
                                              new ServiceImpl(),
                                              bindings.makeRequest(service));
    var sourcePipe = core.createMessagePipe();
    var port = new sample_service.PortPtr();
    var portRequest = bindings.makeRequest(port);

    var foo = new sample_service.Foo();
    foo.bar = new sample_service.Bar();
    foo.name = "Example name";
    foo.source = sourcePipe.handle0;

    assert_equals((await service.frobinate(
        foo, sample_service.Service.BazOptions.REGULAR, port)).result, 42);

    service.ptr.reset();
    serviceBinding.close();

    // sourcePipe.handle1 hasn't been closed yet.
    assert_equals(core.close(sourcePipe.handle1), core.RESULT_OK);

    // portRequest.handle hasn't been closed yet.
    assert_equals(core.close(portRequest.handle), core.RESULT_OK);
  }, 'client server');

  promise_test(async () => {
    var service = new sample_service.ServicePtr();
    // Discard the interface request.
    var interfaceRequest = bindings.makeRequest(service);
    interfaceRequest.close();

    try {
      await service.frobinate(
          null, sample_service.Service.BazOptions.REGULAR, null);
      assert_unreached();
    } catch (e) {
      assert_not_equals(e, null);
    }
  }, 'write to closed pipe');

  promise_test(async () => {
    function ProviderImpl() {
    }

    ProviderImpl.prototype.echoString = function(a) {
      return Promise.resolve({a: a});
    };

    ProviderImpl.prototype.echoStrings = function(a, b) {
      return Promise.resolve({a: a, b: b});
    };

    var provider = new sample_interfaces.ProviderPtr();
    var providerBinding = new bindings.Binding(sample_interfaces.Provider,
                                               new ProviderImpl(),
                                               bindings.makeRequest(provider));
    assert_equals((await provider.echoString("hello")).a, "hello");
    var response = await provider.echoStrings("hello", "world");
    assert_equals(response.a, "hello");
    assert_equals(response.b, "world");

    try {
      core.readMessage = function() {
        return { result: core.RESULT_UNKNOWN };
      };
      await provider.echoString("goodbye");
      assert_unreachable();
    } catch (e) {
      assert_not_equals(e, null);
    }

  }, 'request response');

  done();
});
</script>
